+++ packages/solid-query/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.0...@wundergraph/solid-query@0.4.1) (2023-03-29)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
 ## [0.4.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.3.8...@wundergraph/solid-query@0.4.0) (2023-03-28)
 
 ### Features